Apply updates from the cookiecutter's `db` branch
