+++ .cookiecutter/cookiecutter.json
@@ -5,8 +5,8 @@
   "ignore": [
     "Dockerfile",
     "lms/app.py",
+    "lms/db.py",
     "tests/unit/lms/app_test.py",
-    "lms/pshell.py",
     "lms/migrations/env.py",
     "tests/functional/app_test.py",
     "bin/make_devdata"
