+++ .cookiecutter/includes/requirements/prod.in
@@ -18,11 +18,9 @@ pyramid-googleauth
 pyramid-jinja2
 pyramid-retry
 pyramid-services
-pyramid-tm
 requests
 requests-oauthlib
 sqlparse
 tabulate
 webargs
 xmltodict
-zope.sqlalchemy
